move helpers.ui.js to helpers/ui.js, browserpass/browserpass-extension#320
