+++ src/popup/models/Login.js
@@ -3,7 +3,7 @@
 require("chrome-extension-async");
 const sha1 = require("sha1");
 const helpers = require("../../helpers");
-const helpersUI = require("../../helpers.ui");
+const helpersUI = require("../../helpers/ui");
 const Settings = require("./Settings");
 
 // Search for one of the secret prefixes
